Improve augmentation in ckeditor5-remove-format.
